[GOTY-56] зарефакторил роуты и контроллеры
