+++ src/server/config/index.ts
@@ -1,8 +1,8 @@
 export const config = {
   models: {
     aliases: {
-      UserTheme: 'UserTheme',
-      Theme: 'Theme'
+      UserTheme: 'UserThemeAlias',
+      Theme: 'ThemeAlias'
     }
   }
 }
