+++ src/server/controllers/index.ts
@@ -1,0 +1,4 @@
+export * as authController from './authController'
+export * as themeController from './themeController'
+export * as feedbackController from './feedbackController'
+export * as userController from './userController'
