+++ src/server/routes/feedbackRoutes.ts
@@ -1,22 +1,9 @@
-import express, { Router } from 'express'
-import { Feedback } from '../models'
+import { Router } from 'express'
+
+import { feedbackController } from '../controllers'
 
 const feedbackRouter = Router()
 
-feedbackRouter.post(
-  '/',
-  async (
-    req: express.Request,
-    res: express.Response
-  ): Promise<express.Response> => {
-    try {
-      await Feedback.create(req.body)
-      return res.status(200).send()
-    } catch (err) {
-      console.log(err)
-      return res.status(400).send('There is a problem saving your feedback')
-    }
-  }
-)
+feedbackRouter.post('/', feedbackController.saveFeedback)
 
 export { feedbackRouter }
